Python_GeminiDeveloper_API/requirements.txt
flask
python-dotenv
flask-cors
google-generativeai
supabase
python-dotenv
pytz